Change array type to uint8_t

+++ day03bis.c
@@ -2,12 +2,11 @@
 #include <stdint.h>
 
 uint64_t solve(char t[],int summands){
-    uint64_t T[200]={0};
+    uint8_t T[200]={0};
     int len; for(len=0;t[len]>='0';len++){ T[len]=t[len]-'0'; }
-    int start=0;
-    uint64_t res=0;
+    int start=0; uint64_t res=0;
     for(int k=summands-1;k>=0;k--){
-        unsigned int best=0;   
+        uint8_t best=0;   
         for(int i=start;i<len-k;i++){ if(T[i]>best){start=i+1;best=T[i];} }
         res=10*res+best;
     }
